Create diceClass
Note: the numbers in method names refer to rolls, dice, and sides respectively.  For example, one1d6 is one roll of one six-sided die.  When numbers are replaced by letters (N, X, and Y), the user must determine the number of rolls, dice, and/or sides (so for NXdY, the user inputs the rolls, dice, and sides per die).